feat: allergy dialog 삭제 및 버튼으로 알러지 등록하도록 수정
