+++ src/containers/landing/lunch-menu/allergy/allergy.tsx
@@ -1,71 +1,58 @@
+'use client';
+
 import React from 'react';
-import { Dialog, DialogContent } from '@/components/dialog';
 import { Button } from '@/components/button';
-import { Input } from '@/components/input';
 import { useAllergy } from './allergyContext';
 
-type AllergyDialogProps = {
-  isOpen: boolean;
-  onClose: () => void;
-};
+const allergyList = [
+  { id: 1, name: '난류' },
+  { id: 2, name: '우유' },
+  { id: 3, name: '메밀' },
+  { id: 4, name: '땅콩' },
+  { id: 5, name: '대두' },
+  { id: 6, name: '밀' },
+  { id: 7, name: '고등어' },
+  { id: 8, name: '게' },
+  { id: 9, name: '새우' },
+  { id: 10, name: '돼지고기' },
+  { id: 11, name: '복숭아' },
+  { id: 12, name: '토마토' },
+  { id: 13, name: '아황산류' },
+  { id: 14, name: '호두' },
+  { id: 15, name: '닭고기' },
+  { id: 16, name: '쇠고기' },
+  { id: 17, name: '오징어' },
+  { id: 18, name: '조개류(굴, 전복, 홍합 포함)' },
+  { id: 19, name: '잣' },
+];
 
-function AllergyDialog({ isOpen, onClose }: AllergyDialogProps) {
-  const [allergyInput, setAllergyInput] = React.useState('');
+function AllergyList() {
   const { allergies, setAllergies } = useAllergy();
 
-  const handleAddAllergy = () => {
-    const trimmedAllergy = allergyInput.trim();
-    const allergyArray = trimmedAllergy.split(',').map((item) => item.trim());
-    const newAllergies = allergyArray.filter(
-      (allergy) => allergy && !allergies.includes(allergy),
+  const toggleAllergy = (id: number) => {
+    setAllergies((prev) =>
+      prev.includes(id) ? prev.filter((a) => a !== id) : [...prev, id],
     );
-
-    if (newAllergies.length > 0) {
-      const updatedAllergies = [...allergies, ...newAllergies];
-      setAllergies(updatedAllergies);
-      setAllergyInput('');
-    }
-  };
-
-  const handleRemoveAllergy = (allergy: string) => {
-    const updatedAllergies = allergies.filter((a) => a !== allergy);
-    setAllergies(updatedAllergies);
   };
 
   return (
-    <Dialog open={isOpen} onOpenChange={onClose}>
-      <DialogContent>
-        <span className="text-xm font-semibold mb-4">알러지를 등록하세요.</span>
-        <div className="flex gap-2">
-          <Input
-            value={allergyInput}
-            onChange={(e) => setAllergyInput(e.target.value)}
-            placeholder="알러지 번호 입력 (예: 12, 3, 4)"
-          />
-          <Button onClick={handleAddAllergy}>추가</Button>
-        </div>
-        <div className="mt-4 flex flex-wrap gap-2">
-          {allergies.length > 0 ? (
-            allergies.map((allergy) => (
-              <div
-                key={allergy}
-                className="relative bg-primary-100 text-gray-700 rounded-4 flex items-center justify-center px-4 py-2"
-              >
-                <span
-                  className="text-xl cursor-pointer"
-                  onClick={() => handleRemoveAllergy(allergy)}
-                >
-                  {allergy}
-                </span>
-              </div>
-            ))
-          ) : (
-            <p className="text-gray-500 p-2">등록된 알러지가 없습니다.</p>
-          )}
-        </div>
-      </DialogContent>
-    </Dialog>
+    <div className="p-4">
+      <h3 className="text-sm font-semibold">등록할 알러지를 클릭하세요.</h3>
+      <ul className="flex flex-wrap gap-2 mt-2">
+        {allergyList.map((allergy) => (
+          <li key={allergy.id}>
+            <Button
+              size="sm"
+              className={`text-sm px-2 py-1 rounded-md shadow ${allergies.includes(allergy.id) ? 'bg-primary-400 text-white' : 'bg-gray-200 text-gray-700'}`}
+              onClick={() => toggleAllergy(allergy.id)}
+            >
+              {allergy.id}. {allergy.name} {allergies.includes(allergy.id)}
+            </Button>
+          </li>
+        ))}
+      </ul>
+    </div>
   );
 }
 
-export default AllergyDialog;
+export default AllergyList;
