+++ src/containers/landing/lunch-menu/lunchmenu.tsx
@@ -11,35 +11,12 @@ import useSchoolSearch from './lunchmenu-search/useSchoolSearch';
 import useMealData from './lunchmenu-search/useMealData';
 import SchoolSearchDialog from './lunchmenu-search/schoolSearchDialog';
 import MealList from './mealInformation/mealList';
-import AllergyDialog from './allergy/allergy';
-
-const allergyList = [
-  { id: 1, name: '난류' },
-  { id: 2, name: '우유' },
-  { id: 3, name: '메밀' },
-  { id: 4, name: '땅콩' },
-  { id: 5, name: '대두' },
-  { id: 6, name: '밀' },
-  { id: 7, name: '고등어' },
-  { id: 8, name: '게' },
-  { id: 9, name: '새우' },
-  { id: 10, name: '돼지고기' },
-  { id: 11, name: '복숭아' },
-  { id: 12, name: '토마토' },
-  { id: 13, name: '아황산류' },
-  { id: 14, name: '호두' },
-  { id: 15, name: '닭고기' },
-  { id: 16, name: '쇠고기' },
-  { id: 17, name: '오징어' },
-  { id: 18, name: '조개류(굴, 전복, 홍합 포함)' },
-  { id: 19, name: '잣' },
-];
+import AllergyList from './allergy/allergy';
 
 function LunchMenu() {
   const [schoolName, setSchoolName] = useState<string>('');
   const { schoolList, isLoading, handleSearch } = useSchoolSearch();
   const [isDialogOpen, setIsDialogOpen] = useState(false);
-  const [isAllergyDialogOpen, setIsAllergyDialogOpen] = useState(false);
 
   const [selectedSchool, setSelectedSchool] = useLocalStorage<School | null>(
     'selectedSchool',
@@ -108,30 +85,13 @@ function LunchMenu() {
             >
               학교 등록
             </Button>
-            <Button
-              size="md"
-              variant="primary-ghost"
-              onClick={() => setIsAllergyDialogOpen(true)}
-            >
-              알러지 등록
-            </Button>
           </div>
         }
       />
       <div className="bg-white shadow-custom py-4 rounded-xl w-full overflow-auto">
         {renderContent()}
         <div className="mt-4 p-4">
-          <h3 className="text-sm font-semibold">알러지 정보</h3>
-          <ul className="flex flex-wrap gap-2 mt-2">
-            {allergyList.map((allergy) => (
-              <li
-                key={allergy.id}
-                className="text-sm px-2 py-1 bg-white rounded-md shadow"
-              >
-                {allergy.id}. {allergy.name}
-              </li>
-            ))}
-          </ul>
+          <AllergyList />
         </div>
       </div>
 
@@ -147,10 +107,6 @@ function LunchMenu() {
           setIsDialogOpen(false);
         }}
         handleSearch={handleSearch}
-      />
-      <AllergyDialog
-        isOpen={isAllergyDialogOpen}
-        onClose={() => setIsAllergyDialogOpen(false)}
       />
     </div>
   );
